requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
qrcode[pil]>=7.4.2
pyyaml>=6.0
